--- libs/apifaker/src/main/java/com/woocommerce/android/apifaker/ui/ApiFakerHostFragment.kt
@@ -1,49 +1,0 @@
-package com.woocommerce.android.apifaker.ui
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavType
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import androidx.navigation.navArgument
-import com.woocommerce.android.apifaker.ui.home.HomeScreen
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class ApiFakerHostFragment : Fragment() {
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-
-
-            setContent {
-                val navController = rememberNavController()
-
-                NavHost(
-                    navController = navController,
-                    startDestination = Screen.Home.route()
-                ) {
-                    composable(Screen.Home.route()) {
-                        HomeScreen(hiltViewModel(), navController)
-                    }
-                    composable(
-                        Screen.EndpointDetails.baseRoute,
-                        arguments = listOf(navArgument("endpointId") {
-                            type = NavType.IntType
-                            defaultValue = -1
-                        })
-                    ) {
-                        TODO()
-                    }
-                }
-            }
-        }
-    }
-}
